python310/requirements.txt
oss2~=2.18.4
requests~=2.31.0
Flask~=3.0.3
pillow~=10.3.0
numpy~=1.26.4